Auto-generate infix declarations
