+++ singletons.cabal
@@ -102,6 +102,7 @@ library
                       Data.Singletons.Single.Type,
                       Data.Singletons.Single.Eq,
                       Data.Singletons.Single.Data,
+                      Data.Singletons.Single.Fixity,
                       Data.Singletons.Single,
                       Data.Singletons.TypeLits.Internal,
                       Data.Singletons.Syntax
